worker/requirements.txt
redis
litellm
python-dotenv
google-auth
google-api-python-client
google-cloud
google-cloud-storage